+++ src/features/JsPage/Pages/Page1/Articles/JS_Syntax.js
@@ -1,0 +1,18 @@
+import { 
+    ArticleHeader,
+    SectionHeader,
+    Article, 
+    Section,
+    Highlighted,
+    Hint,
+} from "../../../../../common/Section";
+import CodeSpace from "../../../../CodeSpace/CodeSpace";
+
+export const JS_Syntax = () => (
+<>
+    <ArticleHeader id="4">JavaScript Syntax</ArticleHeader>
+        <Article>
+
+        </Article>
+</>
+);
