+++ src/features/JsPage/Pages/Page1/Articles/JS_Variables.js
@@ -1,0 +1,18 @@
+import { 
+    ArticleHeader,
+    SectionHeader,
+    Article, 
+    Section,
+    Highlighted,
+    Hint,
+} from "../../../../../common/Section";
+import CodeSpace from "../../../../CodeSpace/CodeSpace";
+
+export const JS_Variables = () => (
+<>
+    <ArticleHeader id="5">JavaScript Syntax</ArticleHeader>
+        <Article>
+
+        </Article>
+</>
+);
