+++ src/features/JsPage/Pages/Page2/Page2.js
@@ -1,0 +1,11 @@
+import { ArticleWrapper } from "../../../../common/Section";
+
+export const Page2 = () => (
+    <>
+    <ArticleWrapper>
+
+    </ArticleWrapper>
+    </>
+);
+
+export default Page2;
