+++ src/features/JsPage/index.js
@@ -4,8 +4,8 @@ import NavScrollbar from "../../common/Navigation/NavScrollbar";
 import useThemeColor from "../useThemeColor/useThemeColor";
 import { Outlet } from "react-router-dom";
 import { Route, Routes } from "react-router-dom";
-import Page1 from "./Pages/Page1";
-import Page2 from "./Pages/Page2";
+import Page1 from "./Pages/Page1/JavaScript_Basics";
+import Page2 from "./Pages/Page2/Page2";
 
 const JsPage = () => {
     const themeColor = useThemeColor();
